Broken Loss function so it was not learnign.
I was calculating the loss for 1 sample and using that to train the whole batch, So have to refactor the loss function to account for batchnorm. Also the output layer is not supposed to be normalized but I was normalizing it so I added a simple if statement that skips batchnorm if softmax is applied.